requirements.txt
streamlit
openai
fpdf
python-dotenv
gspread
google-auth
bcrypt
tzdata

